override mmstack parser with is_mmstack=False in Tifffile
